change tag color based off deploy state
